+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [0.1.3](https://github.com/SIMBAChain/libsimba4j/compare/v0.1.3...v0.1.4) (2019-11-28)
+
+
+### Bug Fixes
+
+* Balance amount is now a string
+* transaction hex numbers are properly converted to BigInteger
+
+
+### ⚠ BREAKING CHANGES
+
+* Removed `FullTransaction.getErrorDetails()` method.
+
 
 ## [0.1.3](https://github.com/SIMBAChain/libsimba4j/compare/v0.1.2...v0.1.3) (2019-11-13)